+++ backend/app/tasks/__init__.py
@@ -1,0 +1,5 @@
+from app.tasks.admin_tasks import *
+from app.tasks.post_upload_tasks import *
+from app.tasks.raw_image_processing_tasks import *
+from app.tasks.toolbox_tasks import *
+from app.tasks.upload_tasks import *
