+++ backend/celery/beat-start.sh
@@ -5,4 +5,4 @@ set -o nounset
 
 python /app/app/celeryworker_pre_start.py
 
-celery -A app.tasks beat -s ${CELERY_BEAT_SCHEDULE} -l info
+celery -A app.core.celery_app beat -s ${CELERY_BEAT_SCHEDULE} -l info
