+++ backend/celery/flower-start.sh
@@ -4,7 +4,7 @@ set -o errexit
 set -o nounset
 
 worker_ready() {
-    celery -A app.tasks inspect ping
+    celery -A app.core.celery_app inspect ping
 }
 
 until worker_ready; do
@@ -13,4 +13,4 @@ until worker_ready; do
 done
 echo 'Celery workers are available'
 
-celery -A app.tasks -b "${CELERY_BROKER_URL}" flower
+celery -A app.core.celery_app -b "${CELERY_BROKER_URL}" flower
