+++ backend/celery/worker-start.sh
@@ -5,4 +5,4 @@ set -o nounset
 
 python /app/app/celeryworker_pre_start.py
 
-celery -A app.tasks worker -l info -c 4
+celery -A app.core.celery_app worker -l info -c 4
